requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
distro==1.9.0
fastapi==0.116.1
faiss-cpu==1.12.0
Flask==1.1.1
ffmpy==0.6.0
filelock==3.18.0
fsspec==2025.7.0
gradio==5.38.0
gradio-client==1.11.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
openai==1.97.0
openai-whisper==20250625
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proglog==0.1.12
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
ruff==0.12.4
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==80.9.0
sentence-transformers==5.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
tiktoken==0.9.0
transformers==4.55.4
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
waitress==1.4.3
stripe>=7.0.0
python-multipart